Move shared type guards and assertion functions to root folder
