+++ src/h5web/visualizations/containers/ScalarVisContainer.tsx
@@ -1,7 +1,7 @@
 import React, { ReactElement } from 'react';
 import ScalarVis from '../ScalarVis';
 import { useDatasetValue } from './hooks';
-import { assertDataset } from '../../providers/utils';
+import { assertDataset } from '../../guards';
 import type { VisContainerProps } from './models';
 
 function ScalarVisContainer(props: VisContainerProps): ReactElement {
